Ticket 1032: extracted tests for base writable
