+++ base/uk.ac.stfc.isis.ibex.epics.tests/src/uk/ac/stfc/isis/ibex/epics/tests/writing/SameTypeWritableTest.java
@@ -19,7 +19,7 @@
 
 package uk.ac.stfc.isis.ibex.epics.tests.writing;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 import static org.mockito.Matchers.*;
 import static org.mockito.Mockito.*;
 
@@ -27,7 +27,6 @@ import org.junit.Before;
 import org.junit.Test;
 
 import uk.ac.stfc.isis.ibex.epics.observing.Subscription;
-import uk.ac.stfc.isis.ibex.epics.writing.BaseWritable;
 import uk.ac.stfc.isis.ibex.epics.writing.ConfigurableWriter;
 import uk.ac.stfc.isis.ibex.epics.writing.ForwardingWritable;
 import uk.ac.stfc.isis.ibex.epics.writing.SameTypeWritable;
@@ -39,10 +38,6 @@ public class SameTypeWritableTest {
     private static String VALUE = "123";
 
     private Writable<String> mockDestination;
-    private TestWritable<String> testDestination;
-
-    // This field is used to test functionality of the BaseWritable base class
-    private BaseWritable<String> baseWritable;
 
     // This field is used to test functionality of the ForwardingWritable base
     // class
@@ -53,162 +48,7 @@ public class SameTypeWritableTest {
 
     @Before
     public void setUp() {
-        // Use a mock destination where possible; use a TestWritable when
-        // triggering a method call is needed
         mockDestination = mock(Writable.class);
-        testDestination = new TestWritable<>();
-    }
-
-    @Test
-    public void last_error_is_null_at_initialisation() {
-        // Arrange
-        baseWritable = new SameTypeWritable<String>(mockDestination);
-
-        // Assert
-        assertNull(baseWritable.lastError());
-    }
-
-    @Test
-    public void subscribing_a_writer_updates_the_writer_canWrite() {
-        // Arrange
-        baseWritable = new SameTypeWritable<String>(mockDestination);
-        ConfigurableWriter<String, String> mockWriter = mock(ConfigurableWriter.class);
-
-        // Act
-        baseWritable.subscribe(mockWriter);
-
-        // Assert
-        verify(mockWriter, times(1)).onCanWriteChanged(anyBoolean());
-    }
-
-    @Test
-    public void subscribing_a_writer_updates_the_writer_onError() {
-        // Arrange
-        baseWritable = new SameTypeWritable<String>(mockDestination);
-        ConfigurableWriter<String, String> mockWriter = mock(ConfigurableWriter.class);
-
-        // Act
-        baseWritable.subscribe(mockWriter);
-
-        // Assert
-        verify(mockWriter, times(1)).onError(any(Exception.class));
-    }
-
-    @Test
-    public void last_error_is_updated_on_destination_error() {
-        // Arrange
-        baseWritable = new SameTypeWritable<String>(testDestination);
-
-        Exception exception = new Exception();
-
-        // Act
-        testDestination.simulateError(exception);
-
-        // Assert
-        assertEquals(exception, baseWritable.lastError());
-    }
-
-    @Test
-    public void all_subscribed_writers_are_notified_in_case_of_error() {
-        // Arrange
-        baseWritable = new SameTypeWritable<String>(testDestination);
-
-        ConfigurableWriter<String, String> mockWriter1 = mock(ConfigurableWriter.class);
-        ConfigurableWriter<String, String> mockWriter2 = mock(ConfigurableWriter.class);
-        baseWritable.subscribe(mockWriter1);
-        baseWritable.subscribe(mockWriter2);
-
-        Exception exception = new Exception();
-
-        // Act
-        testDestination.simulateError(exception);
-
-        // Assert
-        verify(mockWriter1, times(1)).onError(exception);
-        verify(mockWriter2, times(1)).onError(exception);
-    }
-
-    @Test
-    public void can_write_is_updated_on_destination_can_write_change_true_case() {
-        // Arrange
-        baseWritable = new SameTypeWritable<String>(testDestination);
-        boolean expected = true;
-
-        // Act
-        testDestination.simulateCanWriteChanged(expected);
-
-        // Assert
-        assertEquals(expected, baseWritable.canWrite());
-    }
-
-    @Test
-    public void can_write_is_updated_on_destination_can_write_change_false_case() {
-        // Arrange
-        baseWritable = new SameTypeWritable<String>(testDestination);
-        boolean expected = false;
-        
-        // Act
-        testDestination.simulateCanWriteChanged(expected);
-
-        // Assert
-        assertEquals(expected, baseWritable.canWrite());
-    }
-
-    @Test
-    public void all_subscribed_writers_are_notified_of_can_write_changed() {
-        // Arrange
-        baseWritable = new SameTypeWritable<String>(testDestination);
-
-        ConfigurableWriter<String, String> mockWriter1 = mock(ConfigurableWriter.class);
-        ConfigurableWriter<String, String> mockWriter2 = mock(ConfigurableWriter.class);
-        baseWritable.subscribe(mockWriter1);
-        baseWritable.subscribe(mockWriter2);
-
-        boolean expected = true;
-
-        // Act
-        testDestination.simulateCanWriteChanged(expected);
-
-        // Assert
-        verify(mockWriter1, times(1)).onCanWriteChanged(expected);
-        verify(mockWriter2, times(1)).onCanWriteChanged(expected);
-    }
-
-    @Test
-    public void subscribing_a_writer_returns_an_unsubscriber_for_the_writer() {
-        // Arrange
-        baseWritable = new SameTypeWritable<String>(testDestination);
-
-        ConfigurableWriter<String, String> mockWriter = mock(ConfigurableWriter.class);
-        Subscription returnedSubscription = baseWritable.subscribe(mockWriter);
-
-        Exception exception = new Exception();
-
-        // Act
-        returnedSubscription.removeObserver();
-
-        // Assert
-        testDestination.simulateError(exception);
-        verify(mockWriter, never()).onError(exception);
-    }
-
-    @Test
-    public void a_writer_gets_subscribed_only_once() {
-        // Arrange
-        baseWritable = new SameTypeWritable<String>(testDestination);
-
-        ConfigurableWriter<String, String> mockWriter = mock(ConfigurableWriter.class);
-
-        Exception exception = new Exception();
-
-        // Act
-        baseWritable.subscribe(mockWriter);
-        baseWritable.subscribe(mockWriter);
-
-        testDestination.simulateError(exception);
-
-        // Assert
-        verify(mockWriter, times(1)).onError(exception);
     }
 
     @Test
